requirements.txt
amqp==5.2.0
ansicon==1.89.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.0
blessed==1.20.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==4.0.4
django-celery-beat==2.6.0
django-debug-toolbar==3.4.0
django-picklefield==3.2
django-timezone-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jinxed==1.2.1
kombu==5.3.7
mccabe==0.7.0
pep8-naming==0.13.2
prompt_toolkit==3.0.46
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.2
pytz==2024.1
redis==5.0.5
requests==2.32.3
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
stripe==9.9.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
